+++ CHANGES.rst
@@ -55,8 +55,9 @@ Cubeviz
 Imviz
 ^^^^^
 
-- ``imviz.default_viewer.save()`` method no longer pops up a file dialog.
-  If the given file exists, it is silently overwritten. [#929]
+- Saving a plot to a PNG (via the astrowidgets API or export plot plugin API) with a provided
+  filename will no longer show the file dialog.  If the given file exists, it is silently
+  overwritten. [#929]
 
 Mosviz
 ^^^^^^